Add PhotonPlanParser unit test
