+++ core/src/test/resources/QualificationExpectations/photon_eventlog_expectation.csv
@@ -1,2 +1,2 @@
 App Name,App ID,Estimated GPU Duration,Estimated GPU Time Saved,SQL DF Duration,SQL Dataframe Task Duration,App Duration,GPU Opportunity,Executor CPU Time Percent,SQL Ids with Failures,Unsupported Read File Formats and Types,Unsupported Write Data Format,Complex Types,Nested Complex Types,Potential Problems,Longest SQL Duration,SQL Stage Durations Sum,NONSQL Task Duration Plus Overhead,Unsupported Task Duration,Supported SQL DF Task Duration,App Duration Estimated,Unsupported Execs,Unsupported Expressions,Estimated Job Frequency (monthly),Total Core Seconds,Photon App
-"Spark shell","local-1629446106683",16424.61,1273.38,1910,6475,17698,1910,27.76,"","","","array<struct<city:string,state:string>>;map<string,map<string,string>>","array<struct<city:string,state:string>>;map<string,map<string,string>>","NESTED COMPLEX TYPE",1453,1203,16292,0,6475,false,"","",30,132,true
+"Databricks Shell","app-20240919162642-0000",205952.38,195501.61,258770,3858136,401454,258770,75.53,"","","","","","",250542,1810943,598848,0,3858136,true,"Execute AddJarsCommand;Execute CreateViewCommand","",30,6213,true
