+++ core/src/test/scala/com/nvidia/spark/rapids/tool/qualification/QualificationSuite.scala
@@ -1758,7 +1758,7 @@ class QualificationSuite extends BaseTestSuite {
   }
 
   test("test support for photon event log") {
-    val logFiles = Array(s"$logDir/photon_eventlog.zstd")  // photon event log
+    val logFiles = Array(s"$logDir/nds_q88_photon_db_13_3.zstd")  // photon event log
     // Status counts: 1 SUCCESS, 0 FAILURE, 0 SKIPPED, 0 UNKNOWN
     val expectedStatus = Some(StatusReportCounts(1, 0, 0, 0))
     runQualificationTest(logFiles, "photon_eventlog_expectation.csv",
